backend/pyproject.toml
[tool.poetry]
name = "christin"
version = "0.1.0"
description = ""
authors = ["harumonia <zxjlm233@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
Flask-Login = "^0.5.0"
Flask-SQLAlchemy = "^2.4.4"
Flask-Migrate = "^2.5.3"
jieba = "^0.42.1"
spacy = "^2.3.5"
loguru = "^0.5.3"
numpy = "1.18.0"
py2neo = "^2020.1.1"
celery = "^5.0.5"
SQLAlchemy = "^1.3.22"
cymysql = "^0.9.15"
flask-security-too = "^4.0.0"
redis = "^3.5.3"
flask-mail = "^0.9.1"
pandas = "^1.2.2"
xlrd = "^2.0.1"
openpyxl = "^3.0.6"
flask-babel = "^2.0.0"
flasgger = "^0.9.5"
PyQRCode = "^1.2.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
flake8 = "^3.8.4"
bandit = "^1.7.0"
pytest-cov = "^2.10.1"
pre-commit = "^2.11.0"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
